refactor: 멘토님 코드리뷰 피드백 적용
